feature/DBCON-319: moves previous tita's tests to new structure
