+++ src/it/tita/src/test/resources/insert-oracle-sys-xmltype-app.xml
