+++ src/it/tita/src/test/resources/stored-procedure-oracle-xmltype-app.xml
